requirements.txt
streamlit
torch
TTS
speechrecognition
pydub
PyPDF2==3.0.0
pytesseract
pillow
